+++ src/utils/minify.js
@@ -9,6 +9,8 @@ export function minify(code) {
       .replaceAll(/transparent/g, "#0000")
       // Remove extra space
       .replaceAll(/\s+/g, " ")
+      // Trim leading 0 before space remove
+      .replaceAll(/(\s)0+\./g, "$1.")
       .replaceAll(/\s*\#/g, "#")
       .replaceAll(/:\s*/g, ":")
       .replaceAll(/,\s*/g, ",")
@@ -29,8 +31,6 @@ export function minify(code) {
       // Remove space between a digit and -
       .replaceAll(/(\d) \-/g, "$1-")
       .replaceAll(/\s*([{}:;,])\s*/g, "$1")
-      // Trim leading 0
-      .replaceAll(/(\D)0+\./g, "$1.")
       // Trim trailing 0
       .replaceAll(/\.(\d*)([1-9])0+(\D)/g, ".$1$2$3")
       .replaceAll(/\+\s+(\S)/g, "+$1")
@@ -50,6 +50,10 @@ export function minify(code) {
       .replaceAll(/\.(\d+)\s+\.(\d)/g, ".$1.$2")
       // Trim space between 2 number with dot and unit: 83Q .5Q => 83Q.5Q
       .replaceAll(/(\d+)([a-zA-Z]+)\s+\.(\d)/g, "$1$2.$3")
+      // Trim space between 1 hexa color and 1 number with dot: #FA1234 .4 => #FA1234.4
+      .replaceAll(/(#[a-fA-F0-9]{6})\s+\.(\d)/g, "$1.$2")
+      .replaceAll(/(#[a-fA-F0-9]{4})\s+\.(\d)/g, "$1.$2")
+      .replaceAll(/(#[a-fA-F0-9]{3})\s+\.(\d)/g, "$1.$2")
       .replaceAll(/([a-z]) (\.\d)/g, "$1$2")
       .replaceAll("&nbsp;", " ")
   );