+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.0
 
 - Leaderboard tool: Add the top 10 leaderboard and your rank
+- Minify: Remove space between an hexa color and a number starting by 0.xx
 
 # 1.1.6